% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pairs.R
\name{get_pairs}
\alias{get_pairs}
\title{Sample two pairs of points from \code{x} and \code{y}}
\usage{
get_pairs(x, y)
}
\arguments{
\item{x}{(numeric) Vector of values from subject 1}

\item{y}{(numeric) Vector of values from subject 2}
}
\value{
(list) List with components \code{x}, the 2 selected values from subject 1, and \code{y}, the 2 selected values from subject 2.
}
\description{
This function random sample two unique pairs of points from the list of
values in \code{x} and \code{y}. Individual indices may be found in the
sample more than once, but pairs are unique.
}
