% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sample.R
\name{make_sample}
\alias{make_sample}
\title{Simulate data having a desired level of ICC}
\usage{
make_sample(
  nSubj,
  nObsPerSubj,
  overall_mean = 0,
  subj_sd = 1,
  obs_sd = 1,
  obs_trans = NULL
)
}
\arguments{
\item{nSubj}{(integer) Number of subjects in the sample}

\item{nObsPerSubj}{(integers) Number of samples per subject. Can be a vector, in which case the number of samples is allowed to vary randomly.}

\item{overall_mean}{(numeric) The overall level of the outcome}

\item{subj_sd}{(numeric) The standard deviation of the subject-level effects}

\item{obs_sd}{(numeric) The standard deviation fo the observation-level effects}

\item{obs_trans}{(function) A unary function to use if the use wants to transform the simulated values (e.g., \code{exp}).}
}
\value{
A \code{data.frame} containing simulated data.
}
\description{
This simulates data from a multi-level normal model. This allows the user to set the overall level of the outcomes and also subject- and observation-level heterogeneity.
}
\examples{
make_sample(10L, 5L, obs_trans = exp)
}
